refactor(rust): Prefix rust module with an underscore to discourage direct usage
